requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
PyJWT==2.8.0

# AI/LLM - Wrapper unificado para múltiplos provedores
litellm>=1.50.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

